feat: refactor listing actions file (#3585)
