+++ sites/partners/src/pages/listings/[id]/index.tsx
@@ -7,7 +7,7 @@ import { ListingStatusBar } from "../../../components/listings/ListingStatusBar"
 import ListingGuard from "../../../components/shared/ListingGuard"
 import { NavigationHeader } from "../../../components/shared/NavigationHeader"
 import Layout from "../../../layouts/index"
-import Aside from "../../../components/listings/Aside"
+import ListingFormActions from "../../../components/listings/ListingFormActions"
 import { ListingContext } from "../../../components/listings/ListingContext"
 import DetailListingData from "../../../components/listings/PaperListingDetails/sections/DetailListingData"
 import DetailListingIntro from "../../../components/listings/PaperListingDetails/sections/DetailListingIntro"
@@ -105,7 +105,7 @@ export default function ListingDetail(props: ListingProps) {
                   </div>
 
                   <div className="w-full md:w-3/12 md:pl-6">
-                    <Aside type="details" />
+                    <ListingFormActions type="details" />
                   </div>
                 </div>
               </div>
